requirements.txt
flask
psycopg2-binary
bcrypt
redis
PyJWT